requirements.txt
absl-py==0.10.0
appdirs==1.4.4
astunparse==1.6.3
attrs==20.2.0
audioread==2.1.8
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
decorator==4.4.2
dill==0.3.3
Flask==1.1.2
future==0.18.2
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
importlib-resources==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kapre==0.3.4
Keras-Preprocessing==1.1.2
librosa==0.8.0
llvmlite==0.34.0
Markdown==3.3
MarkupSafe==1.1.1
numba==0.51.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
Pillow==8.0.1
pooch==1.2.0
promise==2.3
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.14.0
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
SoundFile==0.10.3.post1
spotipy==2.16.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-addons==0.11.2
tensorflow-datasets==4.1.0
tensorflow-estimator==2.3.0
tensorflow-metadata==0.25.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.51.0
typeguard==2.10.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
